Revert to 3D views for SIMD preparation

+++ include/config.h
@@ -1,7 +1,7 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
-using type_real = double;
+using type_real = float;
 const static int ndim{ 2 };
 const static int fint{ 4 }, fdouble{ 8 }, fbool{ 4 }, fchar{ 512 };
 const static bool use_best_location{ true };
